+++ src/dataset_qa_workbench/datasetqaworkbench/main.py
@@ -20,16 +20,20 @@ class DatasetQaWorkbench:
         self.iface = iface
         self.plugin_dir = os.path.dirname(__file__)
         # initialize locale
-        locale = QSettings().value('locale/userLocale')[0:2]
-        locale_path = os.path.join(
-            self.plugin_dir,
-            'i18n',
-            'DatasetQaWorkbench_{}.qm'.format(locale))
+        try:
+            locale = QSettings().value('locale/userLocale')[0:2]
+        except TypeError:
+            pass  # ignore translation stuff
+        else:
+            locale_path = os.path.join(
+                self.plugin_dir,
+                'i18n',
+                'DatasetQaWorkbench_{}.qm'.format(locale))
 
-        if os.path.exists(locale_path):
-            self.translator = QTranslator()
-            self.translator.load(locale_path)
-            QCoreApplication.installTranslator(self.translator)
+            if os.path.exists(locale_path):
+                self.translator = QTranslator()
+                self.translator.load(locale_path)
+                QCoreApplication.installTranslator(self.translator)
 
         # Declare instance attributes
         self.actions = []